# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S10 Window Sticker Generator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface VinInput {
  /**
     * @minLength 17
     * @maxLength 17
     */
  vin: string;
}

export interface S10Option {
  code: string;
  description: string;
  price: number;
}

export interface S10Config {
  year: number;
  trim: string;
  engine: string;
  engineCode?: string;
  engineDisplacement: string;
  engineCylinders: number;
  horsepower: number;
  torque: number;
  fuelType: string;
  transmission: string;
  drivetrain: string;
  cabConfig: string;
  bodyStyle?: string;
  msrpBase: number;
  epaCity: number;
  epaHighway: number;
  standardEquipment: string[];
  availableOptions?: S10Option[];
}

export interface VinDecodeResult {
  vin: string;
  isValid: boolean;
  isS10: boolean;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  make?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  trim?: string | null;
  /** @nullable */
  engineCode?: string | null;
  /** @nullable */
  engineDescription?: string | null;
  /** @nullable */
  transmissionCode?: string | null;
  /** @nullable */
  drivetrainCode?: string | null;
  /** @nullable */
  cabConfig?: string | null;
  /** @nullable */
  bodyStyle?: string | null;
  /** @nullable */
  assemblyPlant?: string | null;
  /** @nullable */
  modelCode?: string | null;
  suggestedConfig?: S10Config | null;
}

export interface StickerOption {
  code: string;
  description: string;
  price: number;
}

export interface StickerInput {
  vin: string;
  year: number;
  make: string;
  model: string;
  trim: string;
  engine: string;
  engineDisplacement?: string;
  /** @nullable */
  engineCylinders?: number | null;
  /** @nullable */
  horsepower?: number | null;
  /** @nullable */
  torque?: number | null;
  transmission: string;
  drivetrain: string;
  cabConfig: string;
  /** @nullable */
  bodyStyle?: string | null;
  /** @nullable */
  assemblyPlant?: string | null;
  /** @nullable */
  exteriorColor?: string | null;
  /** @nullable */
  interiorColor?: string | null;
  msrpBase: number;
  /** @nullable */
  epaCity?: number | null;
  /** @nullable */
  epaHighway?: number | null;
  standardEquipment?: string[];
  selectedOptions?: StickerOption[];
  disclaimerAcknowledged?: boolean;
}

export interface StickerPatch {
  trim?: string;
  engine?: string;
  transmission?: string;
  drivetrain?: string;
  /** @nullable */
  exteriorColor?: string | null;
  /** @nullable */
  interiorColor?: string | null;
  msrpBase?: number;
  /** @nullable */
  epaCity?: number | null;
  /** @nullable */
  epaHighway?: number | null;
  standardEquipment?: string[];
  selectedOptions?: StickerOption[];
  disclaimerAcknowledged?: boolean;
}

export interface Sticker {
  id: number;
  vin: string;
  year: number;
  make: string;
  model: string;
  trim: string;
  engine: string;
  /** @nullable */
  engineDisplacement?: string | null;
  /** @nullable */
  engineCylinders?: number | null;
  /** @nullable */
  horsepower?: number | null;
  /** @nullable */
  torque?: number | null;
  transmission: string;
  drivetrain: string;
  cabConfig: string;
  /** @nullable */
  bodyStyle?: string | null;
  /** @nullable */
  assemblyPlant?: string | null;
  /** @nullable */
  exteriorColor?: string | null;
  /** @nullable */
  interiorColor?: string | null;
  msrpBase: number;
  /** @nullable */
  msrpTotal?: number | null;
  /** @nullable */
  epaCity?: number | null;
  /** @nullable */
  epaHighway?: number | null;
  standardEquipment?: string[];
  selectedOptions?: StickerOption[];
  disclaimerAcknowledged?: boolean;
  /** @nullable */
  shareToken?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface PdfResult {
  url: string;
  filename: string;
}