# lib/api-zod/src/generated/types/pdfResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S10 Window Sticker Generator API
 * OpenAPI spec version: 0.1.0
 */

export interface PdfResult {
  url: string;
  filename: string;
}