# lib/api-zod/src/generated/types/s10Config.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S10 Window Sticker Generator API
 * OpenAPI spec version: 0.1.0
 */
import type { S10Option } from './s10Option';

export interface S10Config {
  year: number;
  trim: string;
  engine: string;
  engineCode?: string;
  engineDisplacement: string;
  engineCylinders: number;
  horsepower: number;
  torque: number;
  fuelType: string;
  transmission: string;
  drivetrain: string;
  cabConfig: string;
  bodyStyle?: string;
  msrpBase: number;
  epaCity: number;
  epaHighway: number;
  standardEquipment: string[];
  availableOptions?: S10Option[];
}